used a risk analysis template for vulns

+++ src/COA-Contracts/Busd.sol
@@ -3,12 +3,12 @@
 pragma solidity ^0.8.4;
 
 // @audit -informational- best pratice to use named imports
-import { ERC20 } from "lib/openzeppelin-contracts/contracts/token/ERC20/ERC20.sol";
+import {ERC20} from "lib/openzeppelin-contracts/contracts/token/ERC20/ERC20.sol";
 
 contract BUSD is ERC20 {
     constructor() ERC20("Busd", "BUSD") {}
 
-    function mint(address _to, uint _amount) external {
+    function mint(address _to, uint256 _amount) external {
         _mint(_to, _amount);
     }
 }
